requirements.txt
requests
newsapi-python